:truck: rename channel store
